REPLACE old API with new API
- The old API caused UB and was not very ergonomic
- Tests are also updated: `cargo test --release`
